% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{check_token_rtoot}
\alias{check_token_rtoot}
\title{check if a token is available and return one if not}
\usage{
check_token_rtoot(token = NULL)
}
\description{
check if a token is available and return one if not
}
