% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{convert_token_to_envvar}
\alias{convert_token_to_envvar}
\title{Convert token to environment variable}
\usage{
convert_token_to_envvar(token, clipboard = TRUE, verbose = TRUE)
}
\arguments{
\item{token}{bearer token, either public or user level}

\item{clipboard}{logical, whether to export the token to the clipboard}

\item{verbose}{logical whether to display messages}
}
\value{
Token (in environment variable format), invisibily
}
\description{
Convert token to environment variable
}
\examples{
\dontrun{
x <- auth_setup("mastodon.social", "public")
envvar <- convert_token_to_envvar(x)
envvar
}
}
