% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_bearer}
\alias{create_bearer}
\title{get a bearer token for the mastodon api}
\usage{
create_bearer(client, type = "public")
}
\arguments{
\item{client}{rtoot client object created with \link{get_client}}

\item{type}{one of "public" or "user". See details}
}
\value{
a mastodon bearer token
}
\description{
get a bearer token for the mastodon api
}
\details{
TBA
}
\references{
https://docs.joinmastodon.org/client/authorized/
}
