% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{create_token}
\alias{create_token}
\title{get a bearer token for the mastodon api}
\usage{
create_token(client, type = "public", browser = TRUE)
}
\arguments{
\item{client}{rtoot client object created with \link{get_client}}

\item{type}{one of "public" or "user". See details}

\item{browser}{if \code{TRUE} (default) a browser window will be opened to authenticate, else the URL will be provided so you can copy/paste this into the browser yourself}
}
\value{
a mastodon bearer token
}
\description{
get a bearer token for the mastodon api
}
\details{
TBA
}
\references{
https://docs.joinmastodon.org/client/authorized/
}
