% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_account}
\alias{get_account}
\title{Query the instance for a specific user}
\usage{
get_account(id, instance = NULL, token = NULL, anonymous = FALSE, parse = TRUE)
}
\arguments{
\item{id}{character, Local ID of a user (this is not the username)}

\item{instance}{character, the server name of the instance where the status is located. If \code{NULL}, the same instance used to obtain the token is used.}

\item{token}{user bearer token (read from file by default)}

\item{anonymous}{some API calls do not need a token. Setting anonymous to TRUE allows to make an anonymous call if possible.}

\item{parse}{logical, if \code{TRUE}, the default, returns a tibble. Use \code{FALSE}  to return the "raw" list corresponding to the JSON returned from the Mastodon API.}
}
\value{
an account
}
\description{
Query the instance for a specific user
}
\examples{
\dontrun{
get_account("109302436954721982")
}
}
