% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_account_favourites}
\alias{get_account_favourites}
\title{Get favourites of user}
\usage{
get_account_favourites(max_id, min_id, limit = 40, token = NULL, parse = TRUE)
}
\arguments{
\item{max_id}{character, Return results older than this id}

\item{min_id}{character, Return results younger than this id}

\item{limit}{integer, maximum number of results to return. Defaults to 40.}

\item{token}{user bearer token}

\item{parse}{logical, if \code{TRUE}, the default, returns a tibble. Use \code{FALSE}  to return the "raw" list corresponding to the JSON returned from the Mastodon API.}
}
\value{
favourited statuses
}
\description{
Get favourites of user
}
\details{
this functions needs a user level auth token
}
