% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_client}
\alias{get_client}
\title{register a mastodon client}
\usage{
get_client(instance = "mastodon.social")
}
\arguments{
\item{instance}{server name}
}
\value{
an rtoot client object
}
\description{
register a mastodon client
}
\references{
https://docs.joinmastodon.org/client/token/#creating-our-application
}
