% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{get_instance}
\alias{get_instance}
\alias{get_instance_general}
\alias{get_instance_peers}
\alias{get_instance_activity}
\alias{get_instance_emoji}
\title{Get various information about a specific instance}
\usage{
get_instance_general(instance = NULL, token = NULL)

get_instance_peers(instance = NULL, token = NULL)

get_instance_activity(instance = NULL, token = NULL)

get_instance_emoji(instance = NULL, token = NULL)
}
\arguments{
\item{instance}{character, the server name of the instance where the status is located. If \code{NULL}, the same instance used to obtain the token is used.}

\item{token}{user bearer token}
}
\value{
list of instance details
}
\description{
Get various information about a specific instance
}
\details{
\describe{
\item{get_instance_general}{Returns general information about the instance}
\item{get_instance_peers}{Returns the peers of an instance}
\item{get_instance_activity}{Shows the weekly activity of the instance (3 months)}
\item{get_instance_emoji}{Lists custom emojis available on the instance}
}
}
