% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.R
\name{get_public_timeline}
\alias{get_public_timeline}
\title{Get the public timeline}
\usage{
get_public_timeline(
  local = FALSE,
  remote = FALSE,
  only_media = FALSE,
  max_id,
  since_id,
  min_id,
  limit = 20L,
  bearer
)
}
\arguments{
\item{local}{logical, Show only local statuses?}

\item{remote}{logical, Show only remote statuses?}

\item{only_media}{logical, Show only statuses with media attached?}

\item{max_id}{character, Return results older than this id}

\item{since_id}{character, Return results newer than this id}

\item{min_id}{character, Return results immediately newer than this id}

\item{limit}{integer, Maximum number of results to return}

\item{bearer}{bearer token from \code{\link[=create_bearer]{create_bearer()}}}
}
\value{
a list of statuses
}
\description{
Query the instance for the public timeline
}
\examples{
bearer <- create_bearer(instance = "social.tchncs.de")
get_public_timeline(bearer = bearer)
}
\references{
https://docs.joinmastodon.org/methods/timelines/
}
