% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token}
\alias{get_token}
\title{get a token for the mastodon api}
\usage{
get_token(instance = "mastodon.social")
}
\arguments{
\item{instance}{server name}
}
\value{
bearer token
}
\description{
get a token for the mastodon api
}
