% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_statuses.R
\name{parse_stream}
\alias{parse_stream}
\title{Parser of Mastodon stream}
\usage{
parse_stream(path)
}
\arguments{
\item{path}{Character, name of JSON file with data collected by any \link{stream_timeline} function.}
}
\description{
Converts Mastodon stream data (JSON file) into a parsed tibble.
}
\details{
The stream sometimes returns invalid lines of json. These are automatically skipped. Parsing can be slow if your json contains a large amount of statuses
}
\examples{
\dontrun{
stream_timeline_public(1,file_name = "stream.json")
parse_stream("stream.json")
}
}
\seealso{
\code{stream_timeline_public()}, \code{stream_timeline_hashtag()},\code{stream_timeline_list()}
}
