% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{post_list_accounts}
\alias{post_list_accounts}
\title{Add accounts to a list}
\usage{
post_list_accounts(id, account_ids, token = NULL, verbose = TRUE)
}
\arguments{
\item{id}{character, id of the list}

\item{account_ids}{ids of accounts to add (this is not the username)}

\item{token}{user bearer token (read from file by default)}

\item{verbose}{logical whether to display messages}
}
\value{
no return value, called for site effects
}
\description{
Add accounts to a list
}
\examples{
\dontrun{
# add some accounts to a list
post_list_create(id = "1234", account_ids = c(1001, 1002))
}
}
