% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{post_list_create}
\alias{post_list_create}
\title{Create a list}
\usage{
post_list_create(title, replies_policy = "list", token = NULL, verbose = TRUE)
}
\arguments{
\item{title}{character, The title of the list to be created}

\item{replies_policy}{character,  One of "followed", "list", or "none". Defaults to "list".}

\item{token}{user bearer token (read from file by default)}

\item{verbose}{logical whether to display messages}
}
\value{
no return value, called for site effects
}
\description{
Create a list
}
\examples{
\dontrun{
# create a list
post_list_create(title = "test")
}
}
