% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.R
\name{post_status}
\alias{post_status}
\title{Perform actions on a status}
\usage{
post_status(id, action = "favourite", token = NULL, verbose = TRUE)
}
\arguments{
\item{id}{character, status id to perform the action on}

\item{action}{character, one of "(un)favourite","(un)reblog","(un)bookmark"}

\item{token}{user bearer token (read from file by default)}

\item{verbose}{logical whether to display messages}
}
\value{
no return value, called for site effects
}
\description{
Perform actions on a status
}
\examples{
\dontrun{
#favourite a status
post_status("xxxxxx",action = "favourite")
#unfavourite a user
post_status("xxxxxx",action = "unfavourite")
}
}
