% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtoot-package.R
\docType{package}
\name{rtoot-package}
\alias{rtoot-package}
\title{rtoot: Collect Mastodon data from R}
\description{
rtoots provides a range of functions to interact with the Mastodon API.
It has five main groups of functions:
\itemize{
\item \verb{get_account_<type>()} to get account level data
\item \verb{get_timeline_<type>()} to get series of statuses
\item \verb{get_instance_<type>()} get information for Mastodon instances
\item \verb{stream_<type>()} to stream statuses from different timeline
\item \verb{post_<type>()} to push content to Mastodon from R
}

(Some auxiliary functions do not follow this naming scheme)

Get started by reading \code{vignette("rtoot")}. and browse all vignettes with \code{browseVignettes(package = "rtoot")}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gesistsa.github.io/rtoot/}
  \item \url{https://github.com/gesistsa/rtoot/}
  \item Report bugs at \url{https://github.com/gesistsa/rtoot/issues}
}

}
\author{
\strong{Maintainer}: David Schoch \email{david@schochastics.net} (\href{https://orcid.org/0000-0003-2952-4812}{ORCID})

Authors:
\itemize{
  \item Chung-hong Chan \email{chainsawtiney@gmail.com} (\href{https://orcid.org/0000-0002-6232-7530}{ORCID})
}

Other contributors:
\itemize{
  \item Johannes Gruber \email{JohannesB.Gruber@gmail.com} (\href{https://orcid.org/0000-0001-9177-1772}{ORCID}) [contributor]
  \item Tim Schatto-Eckrodt \email{kudusch@posteo.de} (\href{https://orcid.org/0000-0003-1658-4373}{ORCID}) [contributor]
}

}
\keyword{internal}
