% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{save_auth_rtoot}
\alias{save_auth_rtoot}
\title{save a bearer token to file}
\usage{
save_auth_rtoot(token, name = NULL, path = NULL)
}
\arguments{
\item{token}{bearer token created with \link{create_token}}

\item{name}{A file name (if you want to store more than one token).}

\item{path}{A path where the token is stored.}
}
\description{
save a bearer token to file
}
