% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{search_accounts}
\alias{search_accounts}
\title{Search the instance for a specific user}
\usage{
search_accounts(
  query,
  limit = 40,
  instance = NULL,
  token = NULL,
  anonymous = FALSE,
  parse = TRUE
)
}
\arguments{
\item{query}{character, search string}

\item{limit}{number of search results to return. Defaults to 40}

\item{instance}{character, the server name of the instance where the status is located. If \code{NULL}, the same instance used to obtain the token is used.}

\item{token}{user bearer token (read from file by default)}

\item{anonymous}{some API calls do not need a token. Setting anonymous to TRUE allows to make an anonymous call if possible.}

\item{parse}{logical, if \code{TRUE}, the default, returns a tibble. Use \code{FALSE}  to return the "raw" list corresponding to the JSON returned from the Mastodon API.}
}
\value{
a tibble ir list of accounts
}
\description{
Search the instance for a specific user
}
\examples{
\dontrun{
search_accounts("schochastics")
}
}
