% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{verify_credentials}
\alias{verify_credentials}
\alias{verify_envvar}
\title{Verify mastodon credentials}
\usage{
verify_credentials(token)

verify_envvar()
}
\arguments{
\item{token}{bearer token, either public or user level}
}
\value{
success or failure message of the verification process
}
\description{
Verify mastodon credentials
}
\details{
If you have created your token as an environment variable, use \code{verify_envvar} to verify your token.
}
\examples{
\dontrun{
#read a token from a file
verify_credentials(token)
}
}
