% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth_setup}
\alias{auth_setup}
\title{Authenticate with a Mastodon instance}
\usage{
auth_setup(instance = NULL, type = NULL, name = NULL)
}
\arguments{
\item{instance}{a public instance of Mastodon (e.g., mastodon.social).}

\item{type}{Either "public" to create a public authentication or "user" to
create authentication for your user (e.g., if you want to post from R or
query your followers).}

\item{name}{give the token a name, in case you want to store more than one.}
}
\value{
A bearer token
}
\description{
Authenticate with a Mastodon instance
}
\examples{
auth_setup("mastodon.social", "public")
}
