% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{get_account_following}
\alias{get_account_following}
\title{Get accounts a user follows}
\usage{
get_account_following(
  id,
  max_id,
  since_id,
  limit = 40,
  token = NULL,
  parse = TRUE
)
}
\arguments{
\item{id}{character, local ID of a user (this is not the username)}

\item{max_id}{character, Return results older than this id}

\item{since_id}{character, Return results newer than this id}

\item{limit}{integer, maximum number of results to return. Defaults to 40.}

\item{token}{user bearer token}

\item{parse}{logical, if \code{TRUE}, the default, returns a tibble. Use \code{FALSE}  to return the "raw" list corresponding to the JSON returned from the Mastodon API.}
}
\value{
followers
}
\description{
Get accounts a user follows
}
\details{
this functions needs a user level auth token
}
