% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{get_fedi_instances}
\alias{get_fedi_instances}
\title{Get a list of fediverse servers}
\usage{
get_fedi_instances(n = 20)
}
\arguments{
\item{n}{number of servers to show}
}
\value{
tibble of fediverse instances
}
\description{
Get a list of fediverse servers
}
\details{
the results are sorted by user count. The function sometimes fails unexpectedly for n>100. Try to rerun your request in that case or use a smaller n.
}
\examples{
\dontrun{
get_fedi_instances(n = 5)
}
}
