% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instances.R
\name{get_fedi_instances}
\alias{get_fedi_instances}
\title{Get a list of fediverse servers}
\usage{
get_fedi_instances(token = "", n = 20, ...)
}
\arguments{
\item{token}{token from instances.social (this is different from your Mastodon token!)}

\item{n}{number of servers to show}

\item{...}{additional parameters for the instances.social API. See \url{https://instances.social/api/doc/}}
}
\value{
tibble of fediverse instances
}
\description{
Get a list of fediverse servers
}
\details{
This function uses the API at instances.social and needs a separate token. Results are sorted by number of users
}
\examples{
\dontrun{
get_fedi_instances(n = 5)
}
}
