% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines.R
\name{get_status}
\alias{get_status}
\title{View specific status}
\usage{
get_status(id, instance = NULL, token = NULL)
}
\arguments{
\item{id}{character, Local ID of a status in the database}

\item{instance}{character, the server name of the instance where the status is located. If \code{NULL}, the same instance used to obtain the token is used.}

\item{token}{user bearer token}
}
\value{
a status
token <- create_bearer(get_client(instance = "social.tchncs.de"), type = "user")
get_status(id = "109298295023649405", token = token)
}
\description{
Query the instance for a specific status
}
