% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{save_auth_rtoot}
\alias{save_auth_rtoot}
\title{save a bearer token to file}
\usage{
save_auth_rtoot(token)
}
\arguments{
\item{token}{bearer token created with \link{create_bearer}}
}
\description{
save a bearer token to file
}
