% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{verify_credentials}
\alias{verify_credentials}
\title{verify mastodon credentials}
\usage{
verify_credentials(token)
}
\arguments{
\item{token}{user bearer token}
}
\value{
mastodon account
}
\description{
verify mastodon credentials
}
